+++ app/src/main/java/com/battlelancer/seriesguide/shows/history/TraktRecentEpisodeHistoryLoader.java
@@ -8,13 +8,13 @@ import androidx.annotation.StringRes;
 import androidx.collection.SparseArrayCompat;
 import com.battlelancer.seriesguide.R;
 import com.battlelancer.seriesguide.SgApp;
-import com.battlelancer.seriesguide.shows.database.SgEpisode2Helper;
 import com.battlelancer.seriesguide.provider.SgRoomDatabase;
-import com.battlelancer.seriesguide.settings.DisplaySettings;
+import com.battlelancer.seriesguide.shows.database.SgEpisode2Helper;
 import com.battlelancer.seriesguide.traktapi.SgTrakt;
 import com.battlelancer.seriesguide.traktapi.TraktCredentials;
 import com.battlelancer.seriesguide.util.Errors;
 import com.battlelancer.seriesguide.util.ImageTools;
+import com.battlelancer.seriesguide.util.LanguageTools;
 import com.battlelancer.seriesguide.util.TextTools;
 import com.battlelancer.seriesguide.util.TimeTools;
 import com.uwetrottmann.androidutils.AndroidUtils;
@@ -123,7 +123,7 @@ public class TraktRecentEpisodeHistoryLoader
             if (showTmdbId != null) {
                 // prefer poster of already added show, fall back to first uploaded poster
                 posterUrl = ImageTools.posterUrlOrResolve(tmdbIdsToPoster.get(showTmdbId),
-                        showTmdbId, DisplaySettings.LANGUAGE_EN, getContext());
+                        showTmdbId, LanguageTools.LANGUAGE_EN, getContext());
             } else {
                 posterUrl = null;
             }