+++ app/src/main/java/com/battlelancer/seriesguide/sync/HexagonShowSync.kt
@@ -10,6 +10,7 @@ import com.battlelancer.seriesguide.shows.database.SgShow2CloudUpdate
 import com.battlelancer.seriesguide.shows.search.discover.SearchResult
 import com.battlelancer.seriesguide.tmdbapi.TmdbTools2
 import com.battlelancer.seriesguide.util.Errors.Companion.logAndReportHexagon
+import com.battlelancer.seriesguide.util.LanguageTools
 import com.github.michaelbull.result.getOrElse
 import com.google.api.client.util.DateTime
 import com.uwetrottmann.androidutils.AndroidUtils
@@ -273,7 +274,7 @@ class HexagonShowSync @Inject constructor(
                 if (!toAdd.containsKey(showTmdbId)) {
                     val item = SearchResult()
                     item.tmdbId = showTmdbId
-                    item.language = show.language
+                    item.language = show.language?.let { LanguageTools.mapLegacyShowCode(it) }
                     item.title = ""
                     toAdd[showTmdbId] = item
                 }