+++ app/src/test/java/com/battlelancer/seriesguide/dataliberation/JsonExportTaskTest.kt
@@ -3,17 +3,17 @@ package com.battlelancer.seriesguide.dataliberation
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import com.battlelancer.seriesguide.EmptyTestApplication
-import com.battlelancer.seriesguide.shows.database.SgEpisode2
 import com.battlelancer.seriesguide.lists.database.SgList
-import com.battlelancer.seriesguide.lists.database.SgListItem
-import com.battlelancer.seriesguide.movies.database.SgMovie
-import com.battlelancer.seriesguide.shows.database.SgSeason2
-import com.battlelancer.seriesguide.shows.database.SgShow2
-import com.battlelancer.seriesguide.movies.database.MovieHelper
-import com.battlelancer.seriesguide.provider.SeriesGuideContract
-import com.battlelancer.seriesguide.shows.database.SgEpisode2Helper
 import com.battlelancer.seriesguide.lists.database.SgListHelper
+import com.battlelancer.seriesguide.lists.database.SgListItem
+import com.battlelancer.seriesguide.movies.database.MovieHelper
+import com.battlelancer.seriesguide.movies.database.SgMovie
+import com.battlelancer.seriesguide.provider.SeriesGuideContract
+import com.battlelancer.seriesguide.shows.database.SgEpisode2
+import com.battlelancer.seriesguide.shows.database.SgEpisode2Helper
+import com.battlelancer.seriesguide.shows.database.SgSeason2
 import com.battlelancer.seriesguide.shows.database.SgSeason2Helper
+import com.battlelancer.seriesguide.shows.database.SgShow2
 import com.battlelancer.seriesguide.shows.database.SgShow2Helper
 import com.battlelancer.seriesguide.shows.episodes.EpisodeFlags
 import com.battlelancer.seriesguide.shows.tools.ShowStatus
@@ -141,7 +141,7 @@ class JsonExportTaskTest {
                 status = ShowStatus.ENDED,
                 poster = "someurl/to/a/poster.jpg",
                 posterSmall = "someurl/to/a/poster.jpg",
-                language = "en",
+                language = "de", // Use language different from default (English).
                 lastUpdatedMs = 0,
                 lastWatchedMs = 1234567890
             ),
@@ -162,7 +162,7 @@ class JsonExportTaskTest {
             )
         )
         const val expectedJsonShows =
-            "[{\"tmdb_id\":95479,\"imdb_id\":\"imdbidvalue\",\"trakt_id\":52,\"title\":\"Jujutsu Kaisen\",\"overview\":\"It\\u0027s all about hollow purple.\",\"language\":\"en\",\"first_aired\":\"2021-02-27T05:11:12.345Z\",\"release_time\":1234,\"release_weekday\":1,\"release_timezone\":\"America/New_York\",\"country\":\"JP\",\"poster\":\"someurl/to/a/poster.jpg\",\"content_rating\":\"\",\"status\":\"ended\",\"runtime\":24,\"genres\":\"Animation|Action \\u0026 Adventure|Sci-Fi \\u0026 Fantasy\",\"network\":\"MBS\",\"rating\":10.0,\"rating_votes\":1234,\"rating_user\":0,\"favorite\":false,\"notify\":true,\"hidden\":false,\"last_watched_ms\":1234567890,\"seasons\":[{\"tmdb_id\":\"1\",\"season\":1,\"episodes\":[{\"tmdb_id\":1,\"episode\":1,\"title\":\"First Episode\",\"first_aired\":1234567890,\"watched\":true,\"plays\":1,\"skipped\":false,\"collected\":false,\"imdb_id\":\"\",\"overview\":\"First overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"},{\"tmdb_id\":2,\"episode\":2,\"title\":\"Second Episode\",\"first_aired\":1234567890,\"watched\":false,\"plays\":0,\"skipped\":true,\"collected\":true,\"imdb_id\":\"\",\"overview\":\"Second overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"}]},{\"tmdb_id\":\"2\",\"season\":2,\"episodes\":[{\"tmdb_id\":1,\"episode\":1,\"title\":\"First Episode\",\"first_aired\":1234567890,\"watched\":true,\"plays\":1,\"skipped\":false,\"collected\":false,\"imdb_id\":\"\",\"overview\":\"First overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"},{\"tmdb_id\":2,\"episode\":2,\"title\":\"Second Episode\",\"first_aired\":1234567890,\"watched\":false,\"plays\":0,\"skipped\":true,\"collected\":true,\"imdb_id\":\"\",\"overview\":\"Second overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"}]}]},{\"imdb_id\":\"\",\"title\":\"\",\"overview\":\"\",\"language\":\"\",\"release_time\":-1,\"release_weekday\":-1,\"release_timezone\":\"\",\"poster\":\"\",\"content_rating\":\"\",\"status\":\"unknown\",\"runtime\":0,\"genres\":\"\",\"network\":\"\",\"rating\":0.0,\"rating_votes\":0,\"rating_user\":0,\"favorite\":false,\"notify\":true,\"hidden\":false,\"last_watched_ms\":0,\"seasons\":[]}]"
+            "[{\"tmdb_id\":95479,\"imdb_id\":\"imdbidvalue\",\"trakt_id\":52,\"title\":\"Jujutsu Kaisen\",\"overview\":\"It\\u0027s all about hollow purple.\",\"language\":\"de\",\"first_aired\":\"2021-02-27T05:11:12.345Z\",\"release_time\":1234,\"release_weekday\":1,\"release_timezone\":\"America/New_York\",\"country\":\"JP\",\"poster\":\"someurl/to/a/poster.jpg\",\"content_rating\":\"\",\"status\":\"ended\",\"runtime\":24,\"genres\":\"Animation|Action \\u0026 Adventure|Sci-Fi \\u0026 Fantasy\",\"network\":\"MBS\",\"rating\":10.0,\"rating_votes\":1234,\"rating_user\":0,\"favorite\":false,\"notify\":true,\"hidden\":false,\"last_watched_ms\":1234567890,\"seasons\":[{\"tmdb_id\":\"1\",\"season\":1,\"episodes\":[{\"tmdb_id\":1,\"episode\":1,\"title\":\"First Episode\",\"first_aired\":1234567890,\"watched\":true,\"plays\":1,\"skipped\":false,\"collected\":false,\"imdb_id\":\"\",\"overview\":\"First overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"},{\"tmdb_id\":2,\"episode\":2,\"title\":\"Second Episode\",\"first_aired\":1234567890,\"watched\":false,\"plays\":0,\"skipped\":true,\"collected\":true,\"imdb_id\":\"\",\"overview\":\"Second overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"}]},{\"tmdb_id\":\"2\",\"season\":2,\"episodes\":[{\"tmdb_id\":1,\"episode\":1,\"title\":\"First Episode\",\"first_aired\":1234567890,\"watched\":true,\"plays\":1,\"skipped\":false,\"collected\":false,\"imdb_id\":\"\",\"overview\":\"First overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"},{\"tmdb_id\":2,\"episode\":2,\"title\":\"Second Episode\",\"first_aired\":1234567890,\"watched\":false,\"plays\":0,\"skipped\":true,\"collected\":true,\"imdb_id\":\"\",\"overview\":\"Second overview\",\"image\":\"/first/still/path.jpg\",\"writers\":\"writers string\",\"gueststars\":\"guest stars string\",\"directors\":\"directors string\"}]}]},{\"imdb_id\":\"\",\"title\":\"\",\"overview\":\"\",\"language\":\"\",\"release_time\":-1,\"release_weekday\":-1,\"release_timezone\":\"\",\"poster\":\"\",\"content_rating\":\"\",\"status\":\"unknown\",\"runtime\":0,\"genres\":\"\",\"network\":\"\",\"rating\":0.0,\"rating_votes\":0,\"rating_user\":0,\"favorite\":false,\"notify\":true,\"hidden\":false,\"last_watched_ms\":0,\"seasons\":[]}]"
 
         val listOfTestSeasons = listOf(
             SgSeason2(